+++ lib/couch_potato/persistence.rb
@@ -12,16 +12,20 @@ require File.dirname(__FILE__) + '/persistence/type_caster'
 require File.dirname(__FILE__) + '/persistence/revisions'
 require File.dirname(__FILE__) + '/forbidden_attributes_protection'
 require File.dirname(__FILE__) + '/view/custom_views'
-require File.dirname(__FILE__) + '/view/lists'
 require File.dirname(__FILE__) + '/view/view_query'
 
 
 module CouchPotato
   module Persistence
+    module TrackModels
+      def inherited(child)
+        super
+        CouchPotato.models << child
+      end
+    end
 
     def self.included(base) #:nodoc:
-      base.send :include, Properties, Callbacks, Json, CouchPotato::View::CustomViews,
-        CouchPotato::View::Lists
+      base.send :include, Properties, Callbacks, Json, CouchPotato::View::CustomViews
       base.send :include, DirtyAttributes, GhostAttributes, Attachments
       base.send :include, MagicTimestamps, ActiveModelCompliance,
         ForbiddenAttributesProtection, Revisions
@@ -31,9 +35,8 @@ module CouchPotato
         alias_method :id, :_id
         alias_method :id=, :_id=
 
-        def self.inherited(child)
-          super
-          CouchPotato.models << child
+        class << self
+          prepend TrackModels
         end
       end