+++ lib/couch_potato/rspec/matchers.rb
@@ -3,7 +3,6 @@ require 'execjs'
 require 'couch_potato/rspec/matchers/map_to_matcher'
 require 'couch_potato/rspec/matchers/reduce_to_matcher'
 require 'couch_potato/rspec/matchers/map_reduce_to_matcher'
-require 'couch_potato/rspec/matchers/list_as_matcher'
 
 module RSpec
   module Matchers
@@ -17,10 +16,6 @@ module RSpec
 
     def rereduce(keys, values)
       CouchPotato::RSpec::ReduceToProxy.new(keys, values, true)
-    end
-
-    def list(results)
-      CouchPotato::RSpec::ListAsProxy.new(results)
     end
 
     def map_reduce(*docs)